This is MERN stack project for demo purpose.
In this a app a user can create an account which has unique email ID.
User can sign-up using the form available in the app or can directly login
using google sign-in. Once the user is signed in, they can create a
reminder using the avilable form. The reminder is displayed on the user
home page which they can modify and delete. the user name is displayed
on the navbar and a total number of reminder is also displayed on their
home page. User can logout using the logout button on the navbar.

The user authentication is done using a middleware logic with
jasonwebtoken in backend which uses node with express. The contoller
logic for user is used for registering the user and signing-in the user
Similarly, the CRUD functionality for the reminders is implemented using
reminder controller logic which is async functions which receive the
data from frontend. The database model is defined using mongoose schema.

The frontend uses React with redux toolkit for the async API calls and
maintaining the global state for user data and reminder.(Tried using
context API and useReducer for implementing global states. The problem
was using multiple context simultaneously ie for user and reminders).
The API calls for CRUD from frontend was implemented using axios methods.

For styling the UI plain CSS along with bootstrap classes and material UI
icons were used.
Routing in the frontend was done using react-router-dom methods.